Introduce PodLifecycleHandlerV0 interface
This breaks out the PodLifecycleHandlerV0 interface. The PodNotifier
interface gives a false sense of what we expect people to implement.

In addition, it marks the PodLifecycleHandlerV0 as deprecated.

Lastly, it breaks out the wrapper code. The benefit of breaking
out the wrapper code is that now people can customize their
reconcilation loop time based on their provider's capabilities.